Implement method for calculating robot head angle relative to direction of movement
Implemented method (tested in "Serpentine Wave Calculator.xlsx") for
calculating the angle of the head of the robot relative to its direction
of movement by calculating the X and Y positions of segments and using
the ratio of the Y vs. servo index linear regression slope to the X vs.
servo index linear regression slope to compute a best-fit slope of the
body position relative to the head, then using the atan2 function to
compute an angle of the head relative to the body segments leading away
from it.