Continue writing documentation; make minor changes and changes based on testing
Continued writing documentation (README.md files and comments in
sketches). The servo-number-to-port-number calculation was moved into
its own function to ease setup, and, for the rectilinear gait, the
"reset" ramp rates are now "positioning" ramp rates and are additionally
used when preparing for turns. Some testing was completed for all three
of these gaits (resulting mainly in keyfob port number changes), though
testing is still not complete for the rectilinear and concertina gaits.