Added optional use of SoftwareSerial port for commands
Added compile-time option to have the sketch use a SoftwareSerial port to receive commands and send responses (currently ports A2 for RX and A3 for TX). Note that these ports conflict with the keyfob's default ports (there are not enough ports on an Arduino Uno to use both simultaneously); a compile-time error was added to alert users of the sketch to this.